InferenceRunner improvements.
