+++ x-pack/plugin/esql/src/main/java/org/elasticsearch/xpack/esql/inference/completion/CompletionOperator.java
@@ -13,7 +13,6 @@ import org.elasticsearch.compute.operator.DriverContext;
 import org.elasticsearch.compute.operator.EvalOperator.ExpressionEvaluator;
 import org.elasticsearch.compute.operator.Operator;
 import org.elasticsearch.core.Releasables;
-import org.elasticsearch.threadpool.ThreadPool;
 import org.elasticsearch.xpack.esql.inference.InferenceOperator;
 import org.elasticsearch.xpack.esql.inference.InferenceRunner;
 import org.elasticsearch.xpack.esql.inference.InferenceRunnerConfig;
@@ -90,7 +89,6 @@ public class CompletionOperator extends InferenceOperator {
      */
     public record Factory(
         InferenceRunner.Factory inferenceRunnerFactory,
-        ThreadPool threadPool,
         String inferenceId,
         ExpressionEvaluator.Factory promptEvaluatorFactory
     ) implements OperatorFactory {
