+++ x-pack/plugin/esql/src/main/java/org/elasticsearch/xpack/esql/planner/LocalExecutionPlanner.java
@@ -61,7 +61,6 @@ import org.elasticsearch.logging.LogManager;
 import org.elasticsearch.logging.Logger;
 import org.elasticsearch.node.Node;
 import org.elasticsearch.tasks.CancellableTask;
-import org.elasticsearch.threadpool.ThreadPool;
 import org.elasticsearch.transport.RemoteClusterAware;
 import org.elasticsearch.xpack.esql.EsqlIllegalArgumentException;
 import org.elasticsearch.xpack.esql.action.ColumnInfoImpl;
@@ -163,7 +162,6 @@ public class LocalExecutionPlanner {
     private final InferenceRunner.Factory inferenceRunnerFactory;
     private final PhysicalOperationProviders physicalOperationProviders;
     private final List<ShardContext> shardContexts;
-    private final ThreadPool threadpool;
 
     public LocalExecutionPlanner(
         String sessionId,
@@ -173,7 +171,6 @@ public class LocalExecutionPlanner {
         BlockFactory blockFactory,
         Settings settings,
         Configuration configuration,
-        ThreadPool threadpool,
         Supplier<ExchangeSource> exchangeSourceSupplier,
         Supplier<ExchangeSink> exchangeSinkSupplier,
         EnrichLookupService enrichLookupService,
@@ -190,7 +187,6 @@ public class LocalExecutionPlanner {
         this.blockFactory = blockFactory;
         this.settings = settings;
         this.configuration = configuration;
-        this.threadpool = threadpool;
         this.exchangeSourceSupplier = exchangeSourceSupplier;
         this.exchangeSinkSupplier = exchangeSinkSupplier;
         this.enrichLookupService = enrichLookupService;
@@ -322,10 +318,7 @@ public class LocalExecutionPlanner {
             source.layout
         );
 
-        return source.with(
-            new CompletionOperator.Factory(inferenceRunnerFactory, threadpool, inferenceId, promptEvaluatorFactory),
-            outputLayout
-        );
+        return source.with(new CompletionOperator.Factory(inferenceRunnerFactory, inferenceId, promptEvaluatorFactory), outputLayout);
     }
 
     private PhysicalOperation planRrfScoreEvalExec(RrfScoreEvalExec rrf, LocalExecutionPlannerContext context) {
@@ -661,7 +654,7 @@ public class LocalExecutionPlanner {
         int scoreChannel = outputLayout.get(rerank.scoreAttribute().id()).channel();
 
         return source.with(
-            new RerankOperator.Factory(inferenceRunnerFactory, threadpool, inferenceId, queryText, rowEncoderFactory, scoreChannel),
+            new RerankOperator.Factory(inferenceRunnerFactory, inferenceId, queryText, rowEncoderFactory, scoreChannel),
             outputLayout
         );
     }
