+++ x-pack/plugin/esql/src/test/java/org/elasticsearch/xpack/esql/inference/rerank/RerankOperatorTests.java
@@ -30,14 +30,7 @@ public class RerankOperatorTests extends InferenceOperatorTestCase<RankedDocsRes
 
     @Override
     protected Operator.OperatorFactory simple(SimpleOptions options) {
-        return new RerankOperator.Factory(
-            mockedInferenceRunnerFactory(),
-            threadPool,
-            SIMPLE_INFERENCE_ID,
-            SIMPLE_QUERY,
-            evaluatorFactory(0),
-            1
-        );
+        return new RerankOperator.Factory(mockedInferenceRunnerFactory(), SIMPLE_INFERENCE_ID, SIMPLE_QUERY, evaluatorFactory(0), 1);
     }
 
     @Override
