+++ x-pack/plugin/esql/src/test/java/org/elasticsearch/xpack/esql/planner/LocalExecutionPlannerTests.java
@@ -318,7 +318,6 @@ public class LocalExecutionPlannerTests extends MapperServiceTestCase {
             null,
             null,
             null,
-            null,
             esPhysicalOperationProviders(shardContexts),
             shardContexts
         );
